check if permission exist
