+++ backend/dataall/core/permissions/api/resolvers.py
@@ -35,7 +35,3 @@ def update_ssm_parameter(context, source, name: str = None, value: str = None):
         parameter_value=value,
     )
     return response
-
-
-def list_invited_organization_permissions_with_descriptions(context, source):
-    return PermissionService.list_invited_organization_permissions_with_descriptions()
