+++ backend/dataall/core/permissions/services/permission_service.py
@@ -7,7 +7,6 @@ from dataall.base.db import exceptions
 from dataall.core.permissions.api.enums import PermissionType
 from dataall.core.permissions.db.permission.permission_models import Permission
 from dataall.core.permissions.db.permission.permission_repositories import PermissionRepository
-from dataall.core.permissions.services.organization_permissions import ORGANIZATION_INVITED_DESCRIPTIONS
 from dataall.core.permissions.services.resources_permissions import RESOURCES_ALL_WITH_DESC
 from dataall.core.permissions.services.tenant_permissions import TENANT_ALL_WITH_DESC
 
@@ -90,10 +89,3 @@ class PermissionService:
             TENANT_ALL_WITH_DESC,
             PermissionType.TENANT,
         )
-
-    @staticmethod
-    def list_invited_organization_permissions_with_descriptions():
-        permissions = []
-        for p in ORGANIZATION_INVITED_DESCRIPTIONS:
-            permissions.append({'name': p, 'description': ORGANIZATION_INVITED_DESCRIPTIONS[p]})
-        return permissions
